+++ python2latex/__init__.py
@@ -4,7 +4,7 @@ from .tex_environment import TexEnvironment, Label
 
 # Other features
 from .document import Document, Section, Subsection
-from .color import Color
+from .color import *
 from .floating_environment import FloatingFigure, FloatingTable, FloatingEnvironmentMixin
 from .plot import Plot, LinePlot, MatrixPlot
 from .template import Template